app/requirements.txt
Flask
paho-mqtt
werkzeug
pytz